Version 1.0
Should be fully functional. I tried making the Move Controller see
inventories that the Interact Controller interacts with but I was unable
to get it working. My progress with this has been commented out in the
InventorySyncHandlers.java if anyone wants to try getting to work. One
of the major road blocks is that InventoryCloseEvent has several issues.
Firstly, it is not cancellable. Second, when exiting an inventory that
was opened with the Player#openInventory() method,
InventoryCloseEvent#getPlayer() will always only return the first
player. Anyways, this version should be fully functional, with minor
cosmetic issues maybe.